new try

+++ hashmap.c
@@ -72,7 +72,7 @@ HashMap * createMap(long capacity) {
     if (map == NULL) return NULL;
     map->buckets = (Pair **)malloc(sizeof(Pair *) * capacity);
 
-    for (long i = 0; i < capacity; i++) {
+    for (int i = 0; i < capacity; i++) {
         map->buckets[i] = NULL;
     }
     map->capacity = capacity;
